+++ sandtrap/__init__.py
@@ -2,7 +2,6 @@ from .errors import StCancelled, StError, StTickLimit, StTimeout, StValidationEr
 from .factory import sandbox
 from .fs import FileSystem, IsolatedFS, VirtualFS
 from .policy import MemberSpec, Policy
-from .refs import find_refs
 from .sandbox import ExecResult
 from .sandbox import Sandbox as Sandbox
 
@@ -18,6 +17,5 @@ __all__ = [
     "StTimeout",
     "StValidationError",
     "VirtualFS",
-    "find_refs",
     "sandbox",
 ]